producer/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
MarkupSafe==1.1.1
protobuf==3.12.2
six==1.15.0
Werkzeug==1.0.1
